test(#192, #177): shared riperf3-test-support crate; retire fixed bind sleeps
#192: new dev-only workspace crate (publish = false) as the single source
for the test-harness pieces that were copy-pasted and had begun to drift:

- free_port -- the #176 sub-ephemeral PID-windowed allocator (was
  byte-identical in riperf3-cli/tests/common and riperf3/tests/common,
  and divergently reimplemented as a bare collision-unchecked 15201+
  counter in riperf3/tests/integration.rs, which could land inside
  another binary's PID window ~1/250);
- udp_serial -- the #191 per-binary UDP serialization lock (statics in
  the rlib stay per-linked-binary, so the semantics are unchanged);
- run_client_with / run_client_ok_with / ClientRun -- the #176
  refused-retry CLI runner (callers pass env!("CARGO_BIN_EXE_riperf3"),
  which only exists while compiling the caller's crate);
- ChildGuard + wait_bounded -- the reaper guard and bounded child wait
  (were private copies in udp_demux.rs / end_block_text.rs).

Both common/mod.rs files become thin re-export/wrapper shims, so no test
call sites change shape.

#177: udp_demux.rs and daemon.rs no longer gate their inline clients on
fixed 300 ms bind sleeps -- the same flake class #176 eliminated
elsewhere. udp_demux wraps its bespoke concurrent-drain client in a
bounded refused-retry loop and captures stderr (was nulled -- a refused
client failed semi-blind); daemon.rs's client goes through run_client,
whose hard timeout still bounds the pre-fix #81 never-serves hang. The
shared `refused()` classifier matches both the pre-#151 Debug rendering
and the iperf3-shaped message, so it is merge-order-proof against the
in-flight #194.

Full workspace green; clippy/fmt clean. Test-only: no product code
touched.
